»Collision 2.0 for object ! Working way better still got to do it for the boundaries, for that we will use the camera, so no touch !

+++ Player.h
@@ -11,7 +11,7 @@ public:
 	Player(Texture::ID id, int nbFrame, int frameRate, const point<int>& srcPos, const point<int>& frameSize);
 	
 	virtual void Move(TileManager* tm) = 0;
-	virtual point<float> GetNextPos(const Vector2D &direction) = 0;
+	virtual point<int> GetNextPos(const Vector2D &direction) = 0;
 
 	virtual ~Player();
 };
